requirements.txt
numpy
Pillow
scipy
matplotlib
scikit-learn
pandas
pdf2image
imagehash
python-poppler